+++ org/postgresql/jdbc4/AbstractJdbc4Clob.java
@@ -3,7 +3,7 @@
 * Copyright (c) 2004-2005, PostgreSQL Global Development Group
 *
 * IDENTIFICATION
-*   $PostgreSQL$
+*   $PostgreSQL: pgjdbc/org/postgresql/jdbc4/AbstractJdbc4Clob.java,v 1.1 2006/06/08 10:34:51 jurka Exp $
 *
 *-------------------------------------------------------------------------
 */
@@ -15,7 +15,7 @@ import java.sql.SQLException;
 public class AbstractJdbc4Clob extends org.postgresql.jdbc3.AbstractJdbc3Clob
 {
 
-    public AbstractJdbc4Clob(org.postgresql.PGConnection conn, int oid) throws SQLException
+    public AbstractJdbc4Clob(org.postgresql.PGConnection conn, long oid) throws SQLException
     {
         super(conn, oid);
     }
